rename script file
