+++ scripts/sword_locales_to_json.js
